openclaw/pyproject.toml
[project]
name = "openclaw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.0,<3.0",    # upper bound — Agno is fast-moving; major versions may break API
    "anthropic>=0.84.0",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
include = ["bot1*", "bot2/src*"]

[tool.pytest.ini_options]
pythonpath = ["bot1", "bot2/src"]
testpaths = ["bot1/tests", "bot2/tests"]
